requirements.txt
streamlit
requests
pandas
python-dotenv
openai
reportlab
google-search-results
duckduckgo-search
